Added float cast for severity score

+++ render.py
@@ -67,7 +67,7 @@ def load_findings_metadata(report_path):
             findings.append(findings_dict)
     # replace Severity by severity_score
     for f in findings:
-        sscore = int(f['severity_score'])
+        sscore = float(f['severity_score'])
         if sscore >= 10:
             f['Severity'] = "Critical"
         elif sscore >= 7 and sscore < 10:
@@ -79,7 +79,7 @@ def load_findings_metadata(report_path):
         elif sscore < 1:
             f['Severity'] = "Info"
     print(findings)
-    findings = sorted(data, key=lambda x: x.get('severity_score', 0))
+    findings = sorted(data, key=lambda x: float(x.get('severity_score', 0)))
     return findings
 
 def compile_pdf(working_path, output_path, utils_path):
